@@ -1,6 +1,8 @@
 module go-hnsw
 
-go 1.22
+go 1.23.0
+
+toolchain go1.23.3
 
 require (
 	github.com/bits-and-blooms/bitset v1.22.0
@@ -10,5 +12,6 @@ require (
 require (
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gonum.org/v1/gonum v0.16.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
